+++ src/GUI/EditLayoutUI.hpp
@@ -12,15 +12,29 @@ namespace qolmod
     {
         protected:
             LayoutModule* module = nullptr;
-            geode::Ref<cocos2d::CCLayer> playerLayer = nullptr;
+            cocos2d::CCLayerColor* selectDarken = nullptr;
             cocos2d::CCDrawNode* previewOutline = nullptr;
+            cocos2d::CCDrawNode* anchorPointsNode = nullptr;
             cocos2d::CCNode* previewNode = nullptr;
+            cocos2d::CCNode* previewNodeParent = nullptr;
             GrabNodeLayer* grab = nullptr;
+            bool lastFrameSelected = false;
+            std::vector<std::pair<float, bool>> anchorPoints = {};
+            float anchorSnapDis = 10;
+            float anchorShowDis = 60;
+            bool snapEnabled = true;
+
+            CCNode* createBackground(bool acu = true);
+
+            void updateOutline();
+            void updateAnchorPoints();
+            void fixToAnchorPoints();
+            void fixToScreenBounds();
 
         public:
             static EditLayoutUI* create(LayoutModule* module);
 
-            void updateOutline();
+            void addAnchorPoint(float position, bool horizontal);
 
             virtual bool setup();
             virtual void update(float dt);